Mirar conexion con datos

+++ src/Semaforo.java
@@ -4,7 +4,7 @@ public class Semaforo {
     short modo; //Color del semaforo 1 verde, 2 amarillo, 4 rojo, 3 apagado
     ConexionExterna
     
-    public Semaforo(int mod) {
+    public Semaforo(short mod) {
         modo = mod;
     }
 
@@ -12,7 +12,7 @@ public class Semaforo {
         modo = mod;
     }
 
-    public int getModo() {
+    public short getModo() {
         return modo;
     }
 }
